+++ lib/stripe_mock/instance.rb
@@ -58,7 +58,7 @@ module StripeMock
     attr_reader :accounts, :balance, :balance_transactions, :bank_tokens, :charges, :coupons, :customers,
                 :disputes, :events, :invoices, :invoice_items, :orders, :payment_intents, :payment_methods,
                 :setup_intents, :plans, :prices, :recipients, :refunds, :transfers, :payouts, :subscriptions, :country_spec,
-                :subscriptions_items, :products, :tax_rates, :checkout_sessions
+                :subscriptions_items, :products, :tax_rates, :checkout_sessions, :checkout_session_line_items
 
     attr_accessor :error_queue, :debug, :conversion_rate, :account_balance
 
@@ -92,6 +92,7 @@ module StripeMock
       @country_spec = {}
       @tax_rates = {}
       @checkout_sessions = {}
+      @checkout_session_line_items = {}
 
       @debug = false
       @error_queue = ErrorQueue.new